+++ plugins/org.eclipse.actf.visualization.gui/src/org/eclipse/actf/accservice/swtbridge/internal/ia2/InternalAccessibleEditableText.java
@@ -22,7 +22,7 @@ public class InternalAccessibleEditableText implements AccessibleEditableText {
 
     private IAccessibleEditableText accessibleEditableText = null;
     
-    public InternalAccessibleEditableText(int address) {
+    public InternalAccessibleEditableText(long address) {
         accessibleEditableText = new IAccessibleEditableText(address);
         accessibleEditableText.AddRef();
     }
