+++ plugins/org.eclipse.actf.visualization.gui/src/org/eclipse/actf/accservice/swtbridge/internal/ia2/InternalAccessibleValue.java
@@ -23,7 +23,7 @@ public class InternalAccessibleValue implements AccessibleValue {
 
     private IAccessibleValue accessibleValue = null;
     
-    public InternalAccessibleValue(int address) {
+    public InternalAccessibleValue(long address) {
         accessibleValue = new IAccessibleValue(address);
         accessibleValue.AddRef();
     }
